:truck: moves handler tests to .../handlers
